fix: 修复追踪数据存储的问题

+++ src/main/java/com/agenthub/api/dto/TraceResponse.java
@@ -23,7 +23,7 @@ public class TraceResponse {
     private Integer statusCode;
     private String errorMessage;
 
-    private Integer totalTokens;
+    private Long totalTokens;
 
     private Instant createdAt;
 }
