+++ src/main/java/com/agenthub/application/dto/TraceOutput.java
@@ -25,7 +25,7 @@ public class TraceOutput {
     private Integer statusCode;
     private String errorMessage;
 
-    private Integer totalTokens;
+    private Long totalTokens;
 
     private Instant createdAt;
 }
